Add routes and methods for retrieving actions.
- Actions get returned as part of a public user profile
- Actions can be retrieved via authenticated GET to /user/actions
- Change PG data type of action meta info to NUMERIC to ensure floats
get stored and retrieved without being manipulated.

+++ db/migrations/20170124130508_CreateActionsTable.sql
@@ -9,9 +9,9 @@ CREATE TABLE IF NOT EXISTS "actions" (
     "user_id"         INTEGER REFERENCES users(user_id),
     "status"          status_options,
     "created_at"      TIMESTAMP,
-    "tons_saved"      REAL,
-    "dollars_saved"   REAL,
-    "upfront_cost"    REAL
+    "tons_saved"      NUMERIC,
+    "dollars_saved"   NUMERIC,
+    "upfront_cost"    NUMERIC
 );
 
 -- +goose Down
